ADX-49 : Isolated bad processingjs compiles from the other artworks, so any code can safely be displayed in the Artwork or Submission lists.
Isolation was achieved by loading each code file from a separate
code.pde endpoint, because ProcessingJS handles errors better when code
is run in this way.

When editing a piece of code, the code is still inlined in the <script>
tag, in order to better facilitate the edit/check process.

Added capabilities to uofa.SeleniumTestCase, to see browser log messages
and filter for errors.

Added tests for the bad code cases.

Updated the README to include information on the selenium test
environment.